Restore links that were accidently erased
On March 14th a task was run from policy publisher that unintentionally
cleared all organisations, lead_organisations, people and working groups
from policies.

To correct, this we are going to replay the event log, ignoring the
commands that caused the problem.

We need to replay more than one event because we now have PatchLinks,
which doesn't require the whole entity.

All except one policy had links populated by a PutContentWithLinks command
in december 2015, so we can ignore anything earlier than that.