Make Specialist Publisher a "protected app"
specialist-publisher is being converted to a "standard" Rails app
that will use the publishing-api V2 endpoints.
We don't know when conversion will be completed and we need
specialist-publisher documents link data in content-store ASAP, hence
this quick hack.

You can revert this commit and remove this exception once specialist-publisher
is fully migrated to use the publishing-api V2 endpoints.