Provide a place to specify reverse link types.
This is to be used by reverse dependencies so the
lookup can follow certain link types backwards.

The same mechanism can be used to specify which
fields to expand in these new link types as is
used in the normal dependency resolution as we are
expecting the reverse of a link type to be unique.