New endpoint for extracting link/content together
This allows us to page through content ids and retrieve all the content items
and links associated with that content id.

For now, we don't apply any filtering (e.g. by publication state). We also
don't want to expand the links hash to include things like base paths, since
this requires more information to uniquely identify the target content item
(such as locale) and we don't actually need this for our usage.

The fields returned for content items are based on the things that influence
content item uniqueness (see doc/object-model-explanation.md), plus
publishing-app and format. We could add more details, but we don't need this
now either.

The data is retrieved using 3 queries. A content ID has many content items and
many links. We query these separately because otherwise we would get a
cartesian product of rows (# of links x # of content items for each content
id).

The pagination works differently to other requests because content id does not
have an inherent ordering. This means that if we define a page by a limit and
and an offset, somebody could publish new content that appears in the middle of
our page between successive queries.

To ensure that the queries links and content items match up to the same set of
content ids, we query all the content ids on the page first, and then pass
those to the other queries.

We are planning to use this in tooling that will check consistency between
rummager and publishing api:
https://github.com/alphagov/finding-things-migration-checker

Content is currently indexed to rummager from multiple sources, and we don't
have a reliable way of rebuilding the index from scratch at the moment. In the
future, content will only be indexed from publishing api.  We want to ensure
that data is kept in sync until then.

We are also in the middle of a data migration from the old tagging
infrastructure (panopticon/content API) to the new one (publishing api/content
store).  Rummager is currently the most complete source of tagging data for
historical content, so comparing publishing API to that gives us a view on any
remaining discrepencies.