Revert "Set document_type of specialist documents to be more specific"
This reverts commit 7611ba79bdd9319682347d743267f4a2b364589d.

We've decided to back out of this data migration as it doesn't amend
the event log to reflect this change. Instead, we'll look at doing
a bulk republish from Specialist Publisher.

This hasn't been deployed to production yet and so data replication
will fix integration's data when it next runs.