Do not validate PatchLinkSet link schema when there is no ContentItem
There is no way to determine which schema to validate against for link
sets without a content_item, unless that information was provided in the
payload.

Until the time where we can add that information to the payload, or
ensure there is a content_item before creating the link set. This will
remove errors that we can't determine which schema to validate against.