Delete HTML publications from publishing api
Due to the various stages of migration of HTML publications, it's best to clear out the partially migrated content items and start from scratch. Example reasons we need to do this:

- Some HTML publications had different content ids across their various editions
- Some HTML publications had their locale set to en when it wasn't, which causes `patch_link_set` calls to send the wrong locale to the Content Store

These are not currently used anywhere in production code. I've included a safety check to prevent this migration being run twice accidentally in production.