Make rcov optional (and enable in jenkins).
Rcov and/or rbenv seem to have a bug which causes
it to take a very long time to do its coverage on
some systems.

We (almost?) never use this coverage locally so
we can disable it everywhere except jenkins or on
the machines of people who are using it for
something (if any).