Move conversion of "format" field into PutContent.
Now that we are no longer storing format in ContentItem, we can treat
that field as a (deprecated) input only; and the sole place where input
creates a ContentItem is in the PutContent command.

Also simplify the validation, now that format is no longer an
alternative to schema_name/document_type.