Delete stale placeholder policies
There are a few placeholder policies which the policy-publisher
doesn't know about.

This is because those policies were renamed at some point before
the migration to publishing-api was completed.

We should delete these manually, since there is no api to do so.
We'll introduce another PR to deal with content-store.