More useful error messages from schema validation.
To allow schemas to accept either the `format` and `document_type`/
`schema_name` fields, they are mostly defined with a single "oneOf"
key at the top level. This results in the schema validator only giving
an unhelpful message that the data did not match one of the schemas,
rather than showing exactly what was wrong.

As a temporary fix, update the SchemaValidator code to extract the
relevant sub-schema based on which field is provided, and promote that
back to the top level to allow it to report the actual problem.