Add last_edited_at to the ContentItem
We need to add a timestamp to content items that records the last time the
content received either a minor or major update. This differs from updated_at
in that it should not update when content is republished or for any other
reason (e.g. when we run a data migration using AR models).