Filter out redirects on `/lookup-by-base-path`
Some base paths have multiple redirects, which are returned
as part of this lookup, but not in any particular order:

```ruby
> ContentItemFilter.filter(state: %w(published unpublished), base_path: ['/renewing-your-tax-credits-claim']).pluck(:base_path, :content_id, :document_typ    e, :updated_at)
[["/renewing-your-tax-credits-claim", "64d2120f-2396-4033-8e91-45f3e1f49ad8", "redirect", Tue, 19 Apr 2016 15:02:19 UTC +00:00],
 ["/renewing-your-tax-credits-claim", "6aecd670-0451-49d2-bce7-a4258be1adb3", "redirect", Tue, 19 Apr 2016 15:02:22 UTC +00:00],
 ["/renewing-your-tax-credits-claim", "bd88b94e-9fb0-44ea-84e8-6c3a4ef962a7", "answer", Tue, 07 Jun 2016 12:27:13 UTC +00:00],
 ["/renewing-your-tax-credits-claim", "fefde62e-a660-41eb-a222-4f711a3208da", "redirect", Wed, 04 May 2016 18:51:16 UTC +00:00]]
 ```

In the example above, the most recently published content item
for this base path is `bd88b94e-9fb0-44ea-84e8-6c3a4ef962a7`. However,
the existing implementation will return the bottom one which is
`fefde62e-a660-41eb-a222-4f711a3208da`, because it's the last value
written to the hash for that base path key.

This commit attempts to address this by filtering out redirects.