Update state of five inconsistent Whitehall documents to 'published'
The Finding Things migration checker has detected an inconsistency in
the state of 5 documents across Whitehall > Publishing API > Content
Store. These items do not exist in the content store, and have
mismatched content IDs in Whitehall/Publishing API.

To fix this, we set the state of these documents to 'published' so that
in Whitehall we can fetch their content IDs via the lookup_content_id
endpoint and write the correct value to Whitehall's database.