Use constants when interpolating SQL.
There was a worry that some of our SQL allowed SQL
injection.  This turned out not to be true, but
there was still uncertainty about the unescaped
string interpolation that occurs in the content
item presenter.

This interpolation is of static SQL strings with
no user input so it's safe, but for the sake of
clarity I've moved these into frozen constants.