Allow full deletion of content including URL.
Content which should never have been on the site
and needs to be fully removed including existence
of its URL (for example, early release of a document
which references a company or stock in its URL)
can send an unpublishing of type "vanish" to
indicate that it was published in error.

This is similar to a "gone" unpublishing except
the URL will 404 instead of 410.