Discard rather than unpublish when substituting drafts
Previously, we transitioned drafts to ‘unpublished’
with a ‘substitute’ type when a substitutable
content item put into the system. This isn’t right
because ‘unpublished’ items are supposed to
reflect the live content store and in this case,
they do not.

Instead, we’ll simply discard drafts when a
substitutable content item is put into the system.
This deletes the previous draft, but we want to
discourage this approach anyway. Ideally,
publishing apps should be using the new /unpublish
endpoint.

If we really had to, we could retrieve the draft
from the event log anyway.