Raise an error if the system is inconsistent
Following a pull request discussion, it was
decided that we should test the assumption that
there is only ever one published/unpublished item
in the history of a content item.

This is a stop-gap before we do this story:
https://trello.com/c/FE1bwOhF/248-validate-content-item-states-across-their-history

It is my opinion that we should remove this
technical debt once the assumption has been
promoted to a constraint as testing assumptions
everywhere litters the code.