Allow publishing apps to request a specific version.
During the Specialist Publisher migration work, we found that we need to be able to retrieve the published version of content so that we can re-trigger email alerts if they fail to send (for whatever reason). Currently, you can only retrieve the latest piece of content, which is sometimes a draft if the content has been redrafted.

Additionally, existing applications provide features that allow users of those apps to view superseded content. Travel Advice Publisher, for example, allows users to view "archived" content which is not currently supported by the Publishing API.