Send validation errors to Errbit as hashes.
The Airbrake/Errbit protocol does not support arrays; anything that is
not a hash is sent as a string. This makes it hard to read the detailed
schema validation errors which are rendered as a single long string.

This extends the pre-processing to convert arrays to hashes keyed by
the array index.