Make version check in put content clearer
I thought I had found a bug in the code where it allowed the state of
the application to be mutated on a validation error. It turns out that
this is actually not the case as all the database changes are wrapped in
a transaction which the exception causes a rollback.

Despite this I thought these tests and the change in the ordering of the
code could make it easier to understand the intentions of the code and
could help someone else not come to the conclusion I came to.