Improve 404 messages for GET /v2/content/:content-id
When a user tries to access a content item which exists but the
requested version or locale is not available they will now receive a 404
message that states that the version and/or locale of that content item
is not found. Previously all 404s returned the same message that implied
there was not a content item for the content_id when it could be a
version or locale not found.