Queries for content item uniqueness
This class provides queries that can be used to determine whether
content items are unique compared to other content items.

A content item is considered unique if the `base_path`,
`locale`, `state` and `user_facing_version` are not the same as those of
another content item.

This was produced to offer a faster alternative to the validations
performed in:
https://github.com/alphagov/publishing-api/blob/cb1c2ac0f5d502c0b4b5a82cd7f653af67c3b238/app/validators/content_item_uniqueness_validator.rb