Remove race condition when discarding a draft where a live item exists
There was a possibility of a race condition with the discard draft command, as
if the Sidekiq job to discard the draft gets run after the job to send the live
item, the content store could end up with nothing at that base path, rather
than the live content item as intended.

This can be avoided by not attempting to delete the draft from the draft
content store, if the live item is being sent.