Use DownstreamDiscardDraftWorker
Switching this command to use the DownstreamDiscardDraftWorker to
discard drafts.

Inserting this in required looking at a couple of issues and putting in
fixes. We can avoid unnecessary discarding (and a potential race
condition) if we only discard when an item is not being replaced.

We also disallow 2 items having the same base_path in the content store,
so we now have appended the locale to the url in the locale test to more
closely match how this works in the application.