Send unpublishings to both content stores
Previously unpublishings were only sent to the live content store, this
sends them to the draft content store as well.

This works on the logic that we don't allow a draft to exist at the time
a content item is unpublished, so immediately after an unpublishing a
content item is in the same state for both content stores.