Conditionally ignore discard draft base path conflicts
The discard draft command is occasionally run from within other
commands. This can result in there being two workers where one is adding
an item to the draft content store and the other removing it - both
potentially at the same base_path.

Normally when the DownstreamDiscardDraftWorker runs it will throw an
error if it is discarding something where there is a base_path
collision, however this change makes this not be logged when the discard
draft is nested. This is to avoid logging when an actual error has not
occurred.