Don't validate uniqueness of pathless content
Content items without a base_path will match anything with a similar locale, version and state.
This will lead to false positives as one of the four uniqueness identifiers is missing.
Instead, we don't attempt to validate pathless items, the existing
version checking in the PutContent protects duplicating pathless
content.