Add VersionForLocaleValidator
This introduces the first validation in the steps towards replacing the
somewhat broken ContentItemUniquenessValidator.

The constraint this validator is concerned with is whether there are
conflicting versions for a content item at a particular locale. To
illustrate consider a record with these attributes:

`(content_id: "5ff5", locale: "en", user_facing_version: 2)`

and the following examples:

1. `(content_id: "5ff5", locale: "en", user_facing_version: 2)`
2. `(content_id: "5ff5", locale: "fr", user_facing_version: 2)`
3. `(content_id: "5ff5", locale: "en", user_facing_version: 3)`
4. `(content_id: "00cc", locale: "en", user_facing_version: 2)`

1. is a conflict as all 3 fields match
2. is not a conflict as we allow 2 content items at the same version to
   exist for different locales
3. is not a conflict as we allow 2 content items with the same locale to
   be different versions
4. is not a conflict as a different content item can be the same locale
   and the same version