Fix publish tests to set document_type
Document type here, is both the value that should be used, and the value that
should be set, as the tests and tested functionality have nothing to do with
validation using schemas.

Also, checking in the event log shows that document_type has the same value as
schema_name for contacts, so the changes that caused this test to break will
not cause any change in behaviour.