Remove PathReservation for '/find-local-council'
This path reservation is a placeholder for the find-local-council
panopticon artefact.  This artefact was created in error and immediately
withdrawn without being published, and without completing the creation of
the associated edition.  There's no workflow for changing a path
reservation to have a different publishing app, so we delete it to make
room for the new one that will come from frontend when it published the
new find-local-council special route.