Restrict Bunny to using a channel only once per message
This changes the way we use channels in Bunny. Previously we were
storing a channel as an instance variable and reusing it for messages
and only closing it if an exception occurred.

Because the queue publisher service was created at the start when rails
initialised it could exist across multiple threads (as far as I can
tell). The docs for Bunny state that "Channels **must not** be shared
between threads. When client publishes a message, at least 2 (typically
3) frames are sent on the wire:".

This commit changes the behaviour so that channels are only created in a
message context ensuring they are not shared across threads.

In initial tests this has appeared to have resolved our `Timeout::Error`
issue that has been plaguing our Errbit on integration, but I'm hestiant
to say it does actually fix it until more examples have run against it.