Add parent_taxons / child_taxons fields to dependent expansion rules
Finding Things are building tools for editing and navigating taxonomies.
To support this, we need to be able to retrieve parts of the taxonomical
hierarchy from the publishing api.

Adding these additional fields to the expansion rules allows us to do
this efficiently, without making multiple network requests to construct
the hierarchy from the client application.