Unpublish substitute return 404 when attempting to reunpublish
This adds a condition (and unfortunately an extra query) when we
unpublish an item to check whether it is unpublished with a type of
"substitute".

It's not really ideal as it could fail if a rogue published document had
a state of published and an unpublishing entry but that seems unlikely.

Rather than getting too bogged down in another substitute exception
clause I have suggested this as an alternative:
https://trello.com/c/EMIigaJ1/306-introduce-a-state-of-substituted#