Refactor unpublish command
Split call method into smaller methods so that it is easier to see what
is happening in the method.

Move content_item into its own method so that it doesn't need to passed
in as an argument everywhere.

Split the validation into 4 separate methods so that it is easier to
read.