Update linkable on publish.
Currently we only update the Linkable in put_content when a new draft
is created with a different base path. This means that if the base
path does not change, the Linkable still refers to the old content
item, so if the title changes the get_linkable endpoint will return
the old data.

This change ensures that on publish any linkable pointing at the old
content_item is deleted and the new one updated. There are three cases:

* if the base_path changed, we have a published linkable pointing at
the old content_item and a draft linkable pointing at the new item;
delete the old one and update the draft one to published.
* if the base_path did not change, we have one linkable but it is
pointing at the old content_item; the delete is a no-op but we update
the linkable to point to the new content_item.
* if there is no previously published item, just update the linkable
state to published.