Remove orphaned translation `ContentItem`
This migration deletes a `ContentItem`, and all the supporting stuff,
that is a translation of a Whitehall `DetailedGuide`. Whitehall has no
knowledge of this translation and the URL currently 404s.

This is probably being caused by a known issue in Whitehall where
deleted translations are not currently unpublished from Publishing API.
There is a [card for
this](https://trello.com/c/ieIgR5mN/692-deleted-translations-not-unpublished) with #core-formats