Accept `override_existing` on /paths endpoint.
This will allow an app to explicitly claim a path that has already been
registered by a different app. Note that this is only supported on this
endpoint; the implicit path reservation that is called as part of the
put_content command will not accept this flag, to make it clear that
this must result from a specific action.