Change semantics of ContentItemPresenter.present
Previously, a query would be performed to lookup the locale of the
provided content item, then the scope would have been restricted to use
that locale. This means that the present method will actually present
the latest content item where the content_id and locale match that of
the provided content item.

I think this is confusing, and redundant, as the only current use of
this method is in the PutContent command, which means that the latest
content item will be the content item that is provided in the method
arguments.

This also saves a database query to determine the locale of the content
item to present.