Enable use of gov.uk external campaign urls[1]
Prior to this commit, external urls were not supported and validation
were put in place to negate them.

This commit enables the use of only gov.uk external campaign urls[1].

Within, external gov.uk campaign urls[1] have been validated with regex:

* possible campaign.service.gov.uk sub domain.
* scheme must either http or https.
* cannot start a hyphen.
* can be alphanumeric (or hyphen).
* number of characters range from 1-63. Refer to [2]

In theory [2], it is believed that DNS itself can have up to 127 levels
of label, each label can be up to 63 characters and the maximum length of
the whole record is limited to approximately 255 characters.

[1] Example: https://new-vat-rates.campaign.gov.uk/path/to/your/new/vat-rates
[2] https://en.wikipedia.org/wiki/Subdomain