Change handling of links to preserve ordering
Currently, the `PATCH link_sets` command works out the difference between the provided links payload and the links currently in the database for the link set concerned. Links are then created and/or deleted in the database to match the payload. This process does not preserve ordering, so new links are appended at the end of the link set.

This commit changes the command to delete all existing links in the database for the link set, and re-add all the links from the payload, thereby preserving the natural ordering in the payload. In addition, the ordering is also recorded in a new `position` field in the `links` table, to allow the ordering algorithm to be changed in future without affecting natural ordering.

All existing links in the database are given a `position` of zero to reflect the fact that they have no inherent order.