Don't delete links for V1 endpoint
https://github.com/alphagov/publishing-api/pull/278 introduced a
"protection" for the links of specialist-publisher. This means that the
V1 endpoint worked exactly the same as V2 for this application.

Unfortunately, specialist-publisher was renamed to manuals-publisher so
our little hack here has stopped working. This causes the links of
manuals to be wiped out on each publishing.

Because manuals-publisher is the only thing sending to the V1 endpoints
anyway, this commit removes all the link-destroying code altogether.