Manage content history with ChangeNote model
We don't want the publishing apps to have to include change history as
part of the content item PUT request any more. Therefore we are moving
to a model whereby we save a ChangeNote for every such request.
Ultimately this will be a top level property in the schemas, but for the
moment we will continue to support `change_note` as part of the details
hash.

Downstream presenter now constructs change history from ChangeNote
records for the content item if they are available.