DownstreamDraftWorker uses content_id and locale
This is instead of a specific content_item_id, this should cause it to
error less as it will be pulling the latest versions from the database.

The errors that could happen before could still happen with this if the
data changes between us loading the data and checking the latest
version. It's important that we still error here to trigger a retry as
otherwise we could lose jobs from the queue.