Remove alert_on_base_path_conflict
This changes how we consider the error that occurs when a draft is
attempted to be discarded but there is an item occupying it's base path.

This situation is no longer considered an error as this is just a
representation that the application state has changed since this job was
queued. Instead it is logged as a warning.