Replace linkables with query
Linkables have proven to be difficult to keep in sync with changes to
state in the system. This removes the code to create and update them and
uses a query to retrieve the same information.

Although not as fast as linkables in some cases, this removes the complication in
attempting to keep the Linkables table up to date. Since it is not as
fast as linkables, we cache the response.

old
topic
..........  0.060000   0.010000   0.070000 (  0.169602)
queries: 12

organisation
..........  0.090000   0.010000   0.100000 (  0.149797)
queries: 10

mainstream_browse_page
..........  0.030000   0.000000   0.030000 (  0.062847)
queries: 10

policy
..........  0.040000   0.010000   0.050000 (  0.159913)
queries: 10

taxon
..........  0.030000   0.000000   0.030000 (  0.034028)
queries: 10

new (warm cache)
topic
..........  0.040000   0.000000   0.040000 (  0.066904)
queries: 10

organisation
..........  0.060000   0.010000   0.070000 (  0.082757)
queries: 10

mainstream_browse_page
..........  0.020000   0.010000   0.030000 (  0.256291)
queries: 10

policy
..........  0.020000   0.000000   0.020000 (  0.058285)
queries: 10

taxon
..........  0.020000   0.010000   0.030000 (  0.180724)
queries: 10