Remove lock version specifying
This block appears to make the tests behave in a non-deterministic
manner as it creates a seconds entry in the lock version table which may
or may not get accessed to present the content item.

It also doesn't seem to add anything to the test or be part of what is
tested.