Update dependent content items by content_id and locale
This changes the way we update dependent content items. Previously we
were pulling out the latest published or draft item for a content item.
Instead we pull out all locales for a content item and enqueue that with
the content id.

This has a secondary advantage that it fixes a potential issue where
content items with multiple locales could have only one instance of
theirs updated when links are updated.