Change autoload_paths to eager_load_paths
When Rails 5 runs in production it does not autoload files anymore.
Any files that are not included within config.eager_load_paths and are
accessed will result in a `NameError: uninitialized constant`.

This problem only affects production environments as both development
and test have the following configurations:

  config.cache_classes = true
  config.eager_load = false

Setting these to false/true respectively allows these issues to be
experienced in development.

Eager load paths are used in autoloading so these do not need to be
defined twice[1].

[1]: https://github.com/rails/rails/blob/v4.1.7/railties/lib/rails/engine.rb#L684