Return a 409 rather than 500 for concurrent content items
For the situation where we receive concurrent put content item requests,
this catches the instance where they fail due to a link set conflict.

This changes the error to be a 409 rather than a 500 which feels more
fitting in this situation.

Note this do not resolve race condition problems in PUT Content.