Don't return `internal_name` from content endpoint
The get content endpoint (/v2/content/:content_id) is currently
returning an `internal_name` attribute. This is calculated from the
`details` hash, with a fallback to the `title` from the root of the
document.

We think it might have been introduced by mistake, because the
`get_linkables` endpoint has this "magic" attribute (it was introduced
in https://github.com/alphagov/publishing-api/pull/216). We assume that
at some point the linkables endpoint used the same presenter as the
content endpoint, causing this functionality to "leak" into another
endpoint.

It has caused some confusion today so we think it's a good idea to
remove it now.

Data in the details hash should be accessed via the details hash, and
there is no need for special behaviour for this attribute.