Lock Content Item during unpublish
We can hit race conditions and deadlocks when concurrent requests occur.
This changes the approach Rails takes when accessing a ContentItem for
editing to use a "SELECT FOR UPDATE" query rather than a normal
"SELECT".

Whilst this transaction is running any other requests for this Content
Item will be locked until this transaction completes.