Switch the header name to a symbol
When provided as a string it is not converted into a header format,
therefore the header is sent as
"govuk_dependency_resolution_source_content_id" rather than
"Govuk-Dependency-Resolution-Source-Content-Id". If it is provided as a
symbol it is converted.

A non standard header name seems to be treated in different ways by
applications, some reject and others convert it to standard format.