A slightly more resilient delete_content_item helper
This no longer uses Linkable which has vanished as a model, and performs
a check whether a table exists on the supporting models for situations
when you are migrating before a table has been created (as was the case
today with change_notes).

This still kinda sucks because of the models in the migration, it'd be
nicer if we had cascading deletes at database level.