Drop Sidekiq log to verbose: false
We're generating a lot of data by having sidekiq logging set to verbose.
On the 26th of October we produced over 20G of log file data where the
vast majority were a log of severity: DEBUG.

I don't believe we're making any use of the data it produces as we
already log in the Rails application the parameters of a Content Item
being sent downstream and if an error occurs we will still get all the
information.