Add public_timestamp when creating ChangeNote
There was an implicit assumption that the payload included a change_note
element in the form of a hash with keys `:note` and `:public_timestamp`.
In fact the schemas just define a string for change_note, so we need to
set the public_timestamp in the Publishing API (which is probably a
better place for it anyway).