Create Actions table migration
The expectation is for items in this table to be looked up by
content_id and locale, as link sets do not concern themselves with a
locale we allow a locale to be nil.

All the relations purposefully are not foreign keys so that the data
they contain can be maintained if the data they reference is removed or
(in the case of events) archived.

Expected columns such as remark and recipient email have not been set up
at this point as there is no workflow for them and adding them
prematurely might complicate them.