Return all appropriate change notes in presenter.
Not all notes will be associated with a content item; however we know
that only historical change notes will be in this state. So we need to
show all notes related to an item with a version up to the current
one, plus all those that have no content_item_id.