Change the way dependency resolution works
Previously, GetDependees was used in dependency resolution. This means
that some of the basic assumed functionality of dependnecy resolution
does not work.

Consider for example two draft content items, A and B, where A has a
link to B. This link means that when A is presented to the content
store, the link to B will be "expanded", including information from B,
e.g. the title. Dependency resolution should trigger A to be presented
to the content store when changes are made to B, to keep the information
in the expanded link consistent with B.

In the terminology of dependency resolution, A is the "dependent" and B
is the "dependee". So, to view the above example in terms of dependency
resolution, upon changes to B (the dependee), any dependents of B must
be presented to the content store, as the information in there expanded
links could have become inconsistent with B.

Previously, GetDependees was used, which means that in the example
above, the dependency resolution worker would look for links from B to
other content items (the dependees). Now, GetDependents is used, which
means that the dependency resolution worker will look for content items
linking to B, fixing the above example.