Rewrite dependency resolution related test
Previously, this was expecting that the dependee was updated, but this
is the wrong way around, as the information that needs updating is in
the dependent content item.

Merge the two expectations together as this makes it simpler, and
through changing the title used in the put content request, explicitly
check that the expanded link is actually updated.