Remove invalid draft content item
This content item has a `draft` state which is not reflected in
Whitehall.
Furthermore it prevents the unpublishing explanation update of the
previous version from Whitehall, because this is not allowed if a draft
is present.
As this draft is not present in Whitehall it is probably inconsistent.
So we decided to remove it, with all its associated supporting objects.