Allow env vars for database configuration
This makes it easier to define different credentials to run the
application. The naming format is based off DATABASE_URL.

I've set 3 environment variables as I don't think there's a way to do it
for dev and test by using a single one as per DATABASE_URL, if anyone
knows an approach let me know.