Refactor ContentDependencies query class
This is a re-approach to ContentDependencies to make it a class that can
act as a one method call approach to access all the content_ids and
locales that require an update as part of dependency resolution. The
major pro for this being that it can be a place that a single set of
tests are wrote to test dependency resolution is working as expected.

Hopefully this has also made the process of dependency resolution easier
to understand.