Make Import command similar to other commands
Get the Import class to inherit from BaseCommand, as the controller expects to
use the call method from BaseCommand.

The BaseCommand defines an initialize method that takes a payload. Switch to
that and remove Import's initialize method.  Because of this, change references
to "params" to "payload".

Add missing response to call method.