Single Queries::LocalesForContentItems query
As we don't use the `for_one` method on Queries::LocalesForContentItem
it makes more sense that we drop this method and only have the
`for_many` one.

However without the context of `for_one` the `for_many` method seems
a confusing name. So Queries::LocalesForContentItem has been renamed to
Queries::LocalesForContentItems and the `for_many` method has been
renamed to use the `call` convention that is across our Query classes
(call is still not a great method name, so suggestions of improvements
welcome ☺️)