Stricter recursion rules in LinkedTo Query
We hit a problem (like one often does with recursion) that a case that
hadn't been considered properly causes the recursion to run a bit crazy.

The problem was caused by allowing matches of any recursive type and
then sorting out whether it matched a path at the end. When this was
used with `ordered_related_items` it was recursing on links between the
various `ordered_related_items` causing a large depth that was run for
every item and eventually killing PostgreSQL.

The remedy for this is to be more strict with how we check for a valid
link path. This has been done by checking that a link path either ends
with a valid sticky link path or has an exact non sticky path.
Unfortunately this has added to the queries complexity.

There are a couple of items in this commit that could be improved:

1) If there is a way to reference the values used in the recursive
select, we could make the query condition significantly simpler
2) I had a nightmare trying to work out whether an array is included
inside another array with the same order. Array equivalent of
"str".include?("st"). Surely Ruby has such a method?