Reaproaching LinkedTo query without recursive SQL
Sadly the recursive SQL had all got a bit complicated and will be
backburned until a time that we need to squeeze that extra performance
now. I'll shed a tear for it later 😥.

This does the recursion via Ruby and only calls a single level items
from the DB.

It still has some complexity working out to correctly obey the "sticky"
rules. Potentially these could be refactored to something simpler
somehow.