Run tests locally using Docker
Builds and tags a “dev” image of the publishing api that tests can be
run against. We boot a postgres container and link to it, so the schema
loading needs to happen on each run. The id for the postgres container
is kept so it can be removed when the script exits.

The app directory is volume mounted into the container so that logs get
saved to the host. The content schemas directory is also mounted into
the container.

This doesn’t run the pact tests yet as I couldn’t work out the external
dependency.