Rigorous popular pillar migration
This updates the populate pillar migration so that once the slow
migrations are run the content item table is locked and anythings that
are changed are caught. The first four updates take between 300-600
seconds each whereas the final updates take about 20s each. So these are
too slow unfortunately to run without a small interruption in service.

This also adds a down migration to repopulate the table if we want to go
back.

I'm editing this migration in a separate commit as it doesn't need to be
run if the previous version of this migration has already been run.