Add content_store column to content_items
This will be used to add a uniqueness constraint to content_items and to
simplify code around knowing which content_store a content_item is in.

We will have a uniqueness constraint on content_store and base_path,
since only one base_path can live in the content_store at any one time.

Secondly it will have a uniqueness constraint on content_store, locale
and content_id to ensure we can't create any duplicate content items.
For example when two requests come in we reject the duplicate at
database level.